+++ .gitignore
@@ -26,3 +26,4 @@ __pycache__
 notification.mp3
 /SwinIR
 /textual_inversion
+/params.json
